+++ src/Resource/Resource.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Jahuty\Resource;
+
+interface Resource
+{
+    public static function from(array $payload);
+}